+++ checkpoint/orbax/checkpoint/_src/asyncio_utils.py
@@ -17,6 +17,8 @@
 import asyncio
 import functools
 from typing import Any, Coroutine, TypeVar
+import nest_asyncio
+
 
 _T = TypeVar('_T')
 
@@ -34,7 +36,15 @@ def as_async_function(func):
   return run
 
 
-def run_sync(coro: Coroutine[Any, Any, _T]) -> _T:
+def run_sync(
+    coro: Coroutine[Any, Any, _T],
+    enable_nest_asyncio: bool = True,  # For testing.
+) -> _T:
   """Runs a coroutine and returns the result."""
-  # TODO: b/324078935 - Support nested asyncio.run calls.
+  try:
+    asyncio.get_running_loop()  # no event loop: ~0.001s, otherwise: ~0.182s
+    if enable_nest_asyncio:
+      nest_asyncio.apply()  # patch asyncio globally in a runtime (idempotent).
+  except RuntimeError:
+    pass
   return asyncio.run(coro)